feat, refactor: block for personal chat, quit and block is only relay to myself and save to my partition
Major chat apps' block doesnt work in group chat, this is smart move and I followed.
So only when personal chat, we check whether sender is blocked or not to receiver when preprocess.
And If sender is blocked, we relay their message only for themselves.
Also block, unblock event message itself is also only sent to the blocker oneself.
Quit event message in both group and personal chat is also only sent to the quitter oneself and this event message will delete local messages,
this follow whatsapp's behavior.
Input will be disable when block checking is not able(so it will be disable also when network failure).
EventMessage display is arranged.
Long Press action is currently disabled when network failure. This can change in future.

+++ backend/messagepreprocess/internal/service/messaging.go
@@ -22,10 +22,30 @@ func (s *Service) ManageMessage(
 	var toIds [][]byte
 	roomId := toId
 	if toIdType == "personal" {
-		toIds = append(toIds, toId[:])
 		toIds = append(toIds, fromId[:])
+		if contentType != "block" && contentType != "unblock" && contentType != "quit" {
+			w, err := s.repository.WasBlocked(ctx, gocql.UUID(toId), gocql.UUID(fromId))
+			if err != nil {
+				return
+			}
+			if !w {
+				toIds = append(toIds, toId[:])
+			}
+		}
 	}
-	if toIdType == "group" && contentType != "create" {
+	if contentType == "block" {
+		err := s.repository.AddBlock(ctx, gocql.UUID(fromId), gocql.UUID(toId))
+		if err != nil {
+			return
+		}
+	}
+	if contentType == "unblock" {
+		err := s.repository.RemoveBlock(ctx, gocql.UUID(fromId), gocql.UUID(toId))
+		if err != nil {
+			return
+		}
+	}
+	if toIdType == "group" && contentType != "create" && contentType != "quit" {
 		participantIds, err := s.repository.FindParticipantIds(ctx, gocql.UUID(toId))
 		if errors.Is(err, gocql.ErrNotFound) {
 			err = nil
@@ -52,11 +72,12 @@ func (s *Service) ManageMessage(
 		}
 		toIds = append(toIds, fromId[:])
 	}
-	if contentType == "quit" {
+	if toIdType == "group" && contentType == "quit" {
 		err := s.repository.RemoveParticipantId(ctx, gocql.UUID(roomId), gocql.UUID(fromId))
 		if err != nil {
 			return
 		}
+		toIds = append(toIds, fromId[:])
 	}
 	if contentType == "image" || contentType == "audio" || contentType == "video" {
 		var wg sync.WaitGroup
